Move local config file utils to configuration package
