+++ java/cdk/src/main/java/sleeper/cdk/Utils.java
@@ -26,10 +26,10 @@ import software.amazon.awscdk.services.logs.LogGroup;
 import software.amazon.awscdk.services.logs.RetentionDays;
 import software.constructs.Construct;
 
+import sleeper.configuration.TablesConfiguration;
 import sleeper.configuration.properties.InstanceProperties;
 import sleeper.configuration.properties.InstanceProperty;
 import sleeper.configuration.properties.table.TableProperties;
-import sleeper.core.schema.Schema;
 
 import java.io.IOException;
 import java.io.Reader;
@@ -38,15 +38,11 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Arrays;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
-import java.util.Objects;
 import java.util.Optional;
-import java.util.Properties;
-import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import static sleeper.configuration.properties.SleeperProperties.loadProperties;
+import static sleeper.configuration.TablesConfiguration.directoryOf;
 import static sleeper.configuration.properties.SystemDefinedInstanceProperty.CONFIG_BUCKET;
 import static sleeper.configuration.properties.UserDefinedInstanceProperty.APACHE_LOGGING_LEVEL;
 import static sleeper.configuration.properties.UserDefinedInstanceProperty.AWS_LOGGING_LEVEL;
@@ -197,65 +193,7 @@ public class Utils {
 
     public static Stream<TableProperties> getAllTableProperties(
             InstanceProperties instanceProperties, Construct scope) {
-        return getAllTableProperties(instanceProperties, getInstancePropertiesPath(scope));
-    }
-
-    public static Stream<TableProperties> getAllTableProperties(
-            InstanceProperties instanceProperties, Path instancePropertiesFile) {
-        Path baseDir = directoryOf(instancePropertiesFile);
-        return streamBaseAndTableFolders(baseDir)
-                .map(folder -> readTablePropertiesFolderOrNull(instanceProperties, folder))
-                .filter(Objects::nonNull);
-    }
-
-    private static TableProperties readTablePropertiesFolderOrNull(
-            InstanceProperties instanceProperties, Path folder) {
-        Path propertiesPath = folder.resolve("table.properties");
-        Path schemaPath = folder.resolve("schema.json");
-        if (!Files.exists(propertiesPath)) {
-            return null;
-        }
-        try {
-            Properties properties = loadProperties(propertiesPath);
-            if (Files.exists(schemaPath)) {
-                Schema schema = Schema.load(schemaPath);
-                return new TableProperties(instanceProperties, schema, properties);
-            }
-            return new TableProperties(instanceProperties, properties);
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    private static Path directoryOf(Path filePath) {
-        Path parent = filePath.getParent();
-        if (parent == null) {
-            return Paths.get(".");
-        } else {
-            return parent;
-        }
-    }
-
-    private static Stream<Path> streamBaseAndTableFolders(Path baseDir) {
-        return Stream.concat(
-                Stream.of(baseDir),
-                streamTableFolders(baseDir));
-    }
-
-    private static Stream<Path> streamTableFolders(Path baseDir) {
-        Path tablesFolder = baseDir.resolve("tables");
-        if (!Files.isDirectory(tablesFolder)) {
-            return Stream.empty();
-        }
-        List<Path> tables;
-        try (Stream<Path> pathStream = Files.list(tablesFolder)) {
-            tables = pathStream
-                    .filter(Files::isDirectory)
-                    .collect(Collectors.toList());
-        } catch (IOException e) {
-            throw new RuntimeException("Failed to list table configuration directories", e);
-        }
-        return tables.stream().sorted();
+        return TablesConfiguration.getAllTableProperties(instanceProperties, getInstancePropertiesPath(scope));
     }
 
     public static void addStackTagIfSet(Stack stack, InstanceProperties properties) {
